requirements.txt
fastapi
uvicorn[standard]
opencv-python-headless
numpy
qrcode[pil]
pillow
